Modified merge runnable

+++ tasks/version-control/merge.js
@@ -1,4 +1,4 @@
-export const description = 'Merges chore, feature, fix or release branches'
+export const description = 'Merges bugfix, chore, feature and hotfix branches into develop and/or release branches'
 
 export const args = {
   delete: {
@@ -11,40 +11,25 @@ export async function run({ delete: deleteBranch }) {
   const { getAllBranches, getCurrentBranch, isCommitRequired, isPushRequired } = await import('./shared.js')
   const currentBranch = await getCurrentBranch()
 
-  if (['develop', 'main'].includes(currentBranch) || ['bugfix', 'chore', 'feature', 'hotfix', 'release'].find((type) => currentBranch.startsWith(`${type}/`))) {
-    // prettier-ignore
-    let targetBranch =
-        currentBranch.startsWith('bugfix') || currentBranch.startsWith('chore') || currentBranch.startsWith('feature') ? 'develop'
-      : currentBranch.startsWith('hotfix') ? 'release'
-      : currentBranch === 'main' ? 'develop'
-      : 'main'
-
-    if (targetBranch === 'release') targetBranch = (await getAllBranches()).local.find((name) => name.startsWith('release'))
+  if (['bugfix', 'chore', 'feature', 'hotfix'].find((type) => currentBranch.startsWith(`${type}/`))) {
+    const targetBranch = currentBranch.startsWith('hotfix') ? (await getAllBranches()).local.find((name) => name.startsWith('release')) : 'develop'
 
     if (targetBranch) {
-      const isRelease = targetBranch === 'main'
       const sequence = []
-      let version
 
       if (await isCommitRequired()) sequence.push('commit -p')
       else if (await isPushRequired()) sequence.push('git push')
 
-      if (isRelease) version = JSON.parse(await (await import('fs/promises')).readFile('package.json')).version
+      sequence.push(`git checkout ${targetBranch}`, 'git pull', `git merge --no-edit -X theirs ${currentBranch}`, 'git push')
 
-      sequence.push(
-        `git checkout ${targetBranch}`,
-        'git pull',
-        `git merge --no-edit ${isRelease ? `-m "Release v${version}"` : '--no-ff'} -X theirs ${currentBranch}`,
-        'git push'
-      )
+      if (targetBranch.startsWith('hotfix')) {
+        sequence.push('git checkout develop', 'git pull', `git merge --no-edit -X theirs ${currentBranch}`, 'git push')
+      }
 
-      if (isRelease) sequence.push(`git tag -a v${version} -m "Version ${version}"`, `git push origin v${version}`)
-
-      if (!['develop', 'main'].includes(currentBranch) && (isRelease || deleteBranch)) {
-        sequence.push(`git branch -D ${currentBranch}`, `git push origin --delete ${currentBranch}`)
-      } else sequence.push(`git checkout ${currentBranch}`)
+      if (deleteBranch) sequence.push(`git branch -D ${currentBranch}`, `git push origin --delete ${currentBranch}`)
+      else sequence.push(`git checkout ${currentBranch}`)
 
       return sequence
     }
-  } else return 'log -e Use merge for the develop or main branches or bugfix, chore, feature, hotfix or release branches'
+  } else return 'log -e Use merge on bugfix, chore, feature or hotfix branches'
 }
